refactor(extract,syntax,hir,linter,analysis,lsp,ide): use u32 for line positions
Update line position types across the codebase:
- Extract crate: Language now from graphql-types, Position uses u32
- Syntax crate: ExtractedBlock.line/character and DocumentRef.line_offset
  now use u32 for LSP compatibility
- HIR crate: BlockContext and structure types use u32 for line_offset
- Linter crate: LintDiagnostic.block_line_offset now u32
- Analysis crate: Cast line offsets appropriately in arithmetic
- LSP crate: Handle u32 line offsets in diagnostic conversion
- IDE crate: Remove unnecessary u32 casts now that types match

This aligns with LSP specification which uses u32 for positions.

+++ crates/analysis/src/lint_integration.rs
@@ -434,7 +434,7 @@ fn convert_lint_diagnostics(
                         offset_end = ld.offset_range.end,
                         start_line_in_block = sl,
                         start_col_in_block = sc,
-                        final_line = sl + block_line_offset,
+                        final_line = sl + block_line_offset as usize,
                         message = %ld.message,
                         "Converting block diagnostic"
                     );
@@ -442,7 +442,7 @@ fn convert_lint_diagnostics(
                 } else {
                     let (sl, sc) = file_line_index.line_col(ld.offset_range.start);
                     let (el, ec) = file_line_index.line_col(ld.offset_range.end);
-                    (0, sl, sc, el, ec)
+                    (0u32, sl, sc, el, ec)
                 };
 
             let severity = match ld.severity {
@@ -456,11 +456,11 @@ fn convert_lint_diagnostics(
                 message: ld.message.into(),
                 range: DiagnosticRange {
                     start: Position {
-                        line: start_line as u32 + line_offset as u32,
+                        line: start_line as u32 + line_offset,
                         character: start_col as u32,
                     },
                     end: Position {
-                        line: end_line as u32 + line_offset as u32,
+                        line: end_line as u32 + line_offset,
                         character: end_col as u32,
                     },
                 },
