+++ crates/ide/src/folding_ranges.rs
@@ -40,7 +40,7 @@ pub fn folding_ranges(
 
     for doc in parse.documents() {
         let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-        let line_offset = doc.line_offset as u32;
+        let line_offset = doc.line_offset;
 
         let doc_cst = doc.tree.document();
